+++ mixins/highlighted_stroke.mixin.js
