+++ shape_circle.js
@@ -35,10 +35,10 @@ var Circle = Fabric.util.createClass(Fabric.Circle, {
 });
 
 var proto = Circle.prototype;
-mixin(proto, require('./clone.mixin'));
-mixin(proto, require('./highlighted_stroke.mixin'));
-mixin(proto, require('./limit_size'));
-mixin(proto, require('./nudge'));
+mixin(proto, require('./mixins/clone.mixin'));
+mixin(proto, require('./mixins/highlighted_stroke.mixin'));
+mixin(proto, require('./mixins/limit_size'));
+mixin(proto, require('./mixins/nudge'));
 
 /**
  * Increment the size of the circle about its center.
