+++ shape_line.js
@@ -17,11 +17,11 @@ var Line = Fabric.util.createClass(Fabric.Line, {
 });
 
 var proto = Line.prototype;
-mixin(proto, require('./clone.mixin'));
-mixin(proto, require('./highlighted_stroke.mixin'));
-mixin(proto, require('./limit_size'));
-mixin(proto, require('./nudge'));
-mixin(proto, require('./two_point_interactivity'));
+mixin(proto, require('./mixins/clone.mixin'));
+mixin(proto, require('./mixins/highlighted_stroke.mixin'));
+mixin(proto, require('./mixins/limit_size'));
+mixin(proto, require('./mixins/nudge'));
+mixin(proto, require('./mixins/two_point_interactivity'));
 
 // Overrides the method from the to_object.mixin beccause the
 // constructor of Line is differen than the rest of the shapes.
